requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap5==2024.2
cryptography==43.0.3
Django==5.0.4
django-allauth==64.0.0
django-crispy-forms==2.3
google-api-core==2.23.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyodbc==5.1.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.1
razorpay==1.4.2
requests==2.32.3
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
